All the modifications to auton from Thursday at states
including
adding cross field auton - working
dual cube auton - still needs testing but doesn't affect anythings else
distance changes to scale auton
changed the ending for DriveDistance to look at the the encoder that has
the bigger value, not the average - at the request of Mr. T and Pete T

Also, because of the gear box switch before forest hills, forward is now
positive on the encoders. I fixed this by multiplying the distance by -1
in the constructor of DriveDistance so we wouldn't have to change
everywhere else in the code from a negative distance to a positive
distance.